+++ config/detekt/detekt-baseline.xml
@@ -6,9 +6,13 @@
     <ID>EmptyCatchBlock:AdvertisingIdClient.kt$AdvertisingIdClient.AdvertisingConnection${ }</ID>
     <ID>EmptyCatchBlock:Purchases.kt$Purchases.Companion.&lt;no name provided>${ }</ID>
     <ID>EmptyFunctionBlock:AdvertisingIdClient.kt$AdvertisingIdClient.AdvertisingConnection${}</ID>
+    <ID>LargeClass:BillingWrapperTest.kt$BillingWrapperTest</ID>
     <ID>LargeClass:Purchases.kt$Purchases : LifecycleDelegate</ID>
     <ID>LongMethod:Purchases.kt$Purchases$ fun restorePurchases( callback: ReceiveCustomerInfoCallback )</ID>
+    <ID>LongParameterList:BillingWrapperTest.kt$BillingWrapperTest$( sku: String, purchaseToken: String, productType: ProductType, offeringIdentifier: String? = null, purchaseState: Int = Purchase.PurchaseState.PURCHASED, acknowledged: Boolean = false )</ID>
     <ID>LongParameterList:StubGooglePurchase.kt$( productIds: List&lt;String> = listOf("com.revenuecat.lifetime"), purchaseTime: Long = System.currentTimeMillis(), purchaseToken: String = "abcdefghijipehfnbbnldmai.AO-J1OxqriTepvB7suzlIhxqPIveA0IHtX9amMedK0KK9CsO0S3Zk5H6gdwvV" + "7HzZIJeTzqkY4okyVk8XKTmK1WZKAKSNTKop4dgwSmFnLWsCxYbahUmADg", signature: String = "signature${System.currentTimeMillis()}", purchaseState: Int = Purchase.PurchaseState.PURCHASED, acknowledged: Boolean = true, orderId: String = "GPA.3372-4150-8203-17209" )</ID>
+    <ID>MagicNumber:BillingResponseCodeTest.kt$BillingResponseCodeTest$12</ID>
+    <ID>MagicNumber:BillingWrapperTest.kt$BillingWrapperTest$200</ID>
     <ID>MagicNumber:SampleWeatherData.kt$SampleWeatherData.Companion$120</ID>
     <ID>MagicNumber:SampleWeatherData.kt$SampleWeatherData.Companion$20</ID>
     <ID>MagicNumber:SampleWeatherData.kt$SampleWeatherData.Companion$32</ID>
@@ -18,27 +22,6 @@
     <ID>MagicNumber:SampleWeatherData.kt$SampleWeatherData.Companion$90</ID>
     <ID>MagicNumber:SampleWeatherData.kt$SampleWeatherData.Companion$91</ID>
     <ID>MagicNumber:StubGooglePurchase.kt$4</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendBadRequest$7226</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendCannotTransferPurchase$7102</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendEmptyAppUserId$7220</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInternalServerError$7110</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAPIKey$7225</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAppStoreSharedSecret$7104</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAppleSubscriptionKey$7234</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAuthToken$7224</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidPaymentModeOrIntroPriceNotProvided$7105</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidPlatform$7000</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidPlayStoreCredentials$7107</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidReceiptToken$7103</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidSubscriberAttributes$7263</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidSubscriberAttributesBody$7264</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendPlayStoreGenericError$7231</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendPlayStoreInvalidPackageName$7230</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendPlayStoreQuotaExceeded$7229</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendProductIDsMalformed$7662</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendProductIdForGoogleReceiptNotProvided$7106</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendStoreProblem$7101</ID>
-    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendUserIneligibleForPromoOffer$7232</ID>
     <ID>MaxLineLength:Constants.kt$Constants$/* The entitlement ID from the RevenueCat dashboard that is activated upon successful in-app purchase for the duration of the purchase. Modify this property to reflect your app's entitlement identifier. */</ID>
     <ID>NoWildcardImports:com.revenuecat.sample.ui.paywall.PaywallFragment.kt:10</ID>
     <ID>NoWildcardImports:com.revenuecat.sample.ui.user.UserFragment.kt:14</ID>
@@ -46,7 +29,6 @@
     <ID>SwallowedException:DeviceCache.kt$DeviceCache$catch (e: ClassCastException) { emptySet() }</ID>
     <ID>SwallowedException:DeviceCache.kt$DeviceCache$catch (e: JSONException) { null }</ID>
     <ID>SwallowedException:DeviceCache.kt$DeviceCache$catch (e: NullPointerException) { emptySet() }</ID>
-    <ID>SwallowedException:HTTPClient.kt$HTTPClient$catch (e: IOException) { connection.errorStream }</ID>
     <ID>SwallowedException:Purchases.kt$Purchases.Companion.&lt;no name provided>$catch (e: IllegalArgumentException) { // Play Services not available callback.onReceived(false) }</ID>
     <ID>SwallowedException:Purchases.kt$Purchases.Companion.&lt;no name provided>$catch (e: IllegalArgumentException) { }</ID>
     <ID>ThrowingExceptionsWithoutMessageOrCause:AdvertisingIdClient.kt$AdvertisingIdClient.AdvertisingConnection$IllegalStateException()</ID>
@@ -56,6 +38,7 @@
     <ID>TooManyFunctions:AmazonBilling.kt$AmazonBilling : BillingAbstractProductDataResponseListenerPurchaseResponseListenerPurchaseUpdatesResponseListenerUserDataResponseListener</ID>
     <ID>TooManyFunctions:Backend.kt$Backend</ID>
     <ID>TooManyFunctions:BillingWrapper.kt$BillingWrapper : BillingAbstractPurchasesUpdatedListenerBillingClientStateListener</ID>
+    <ID>TooManyFunctions:BillingWrapperTest.kt$BillingWrapperTest</ID>
     <ID>TooManyFunctions:DeviceCache.kt$DeviceCache</ID>
     <ID>TooManyFunctions:HTTPClient.kt$HTTPClient</ID>
     <ID>TooManyFunctions:Purchases.kt$Purchases : LifecycleDelegate</ID>
